[REFACTOR] localization middleware
