+++ backend/helpers/helpers.go
@@ -1,5 +1,11 @@
 package helpers
 
+import (
+	"path/filepath"
+	"runtime"
+	"strings"
+)
+
 func Contains(s []string, str string) bool {
 	for _, v := range s {
 		if v == str {
@@ -8,4 +14,13 @@ func Contains(s []string, str string) bool {
 	}
 
 	return false
+}
+
+func GetRootPath() string {
+	_, b, _, _ := runtime.Caller(0)
+
+	// Root folder of this project
+	root := strings.ReplaceAll(filepath.Join(filepath.Dir(b), ".."), `\`, `/`)
+
+	return root
 }
